refactor (all): move default kpi to shared folder
